+++ demo/CheckBoxDemo.qml
@@ -40,13 +40,13 @@ ColumnLayout {
                     color: index == 0 ? Theme.light.textColor : Theme.dark.textColor
                 }
 
-                Checkbox {
+                CheckBox {
                     checked: true
                     text: "On"
                     darkBackground: index == 1
                 }
 
-                Checkbox {
+                CheckBox {
                     checked: true
                     enabled: false
                     text: "Disabled"
@@ -58,12 +58,12 @@ ColumnLayout {
                     color: index == 0 ? Theme.light.textColor : Theme.dark.textColor
                 }
 
-                Checkbox {
+                CheckBox {
                     text: "Off"
                     darkBackground: index == 1
                 }
 
-                Checkbox {
+                CheckBox {
                     text: "Disabled"
                     enabled: false
                     darkBackground: index == 1