+++ demo/main.qml
@@ -6,12 +6,14 @@ ApplicationWindow {
     id: demo
 
     theme {
-        accentColor: "#009688"
+        primaryColor: Palette.colors["blue"]["500"]
+        primaryDarkColor: Palette.colors["blue"]["700"]
+        accentColor: Palette.colors["teal"]["500"]
     }
 
     property var components: [
-            "Button", "Checkbox", "Color Palette", "Dialog", "Forms", "Icon", "List Items", "Page Stack",
-            "Progress Bar", "Radio Button", "Slider", "Switch", "TextField"
+            "Button", "CheckBox", "Color Palette", "Dialog", "Forms", "Icon", "List Items", "Page Stack",
+            "Progress Bar", "Radio Button", "Slider", "Switch", "TextField", "Typography"
     ]
 
     property string selectedComponent: components[0]